requirements.txt
streamlit
pdfplumber
python-docx
pandas
openai>=1.0.0
